Trivial: move repository and blockstore to client package
