+++ src/test/java/org/ethereum/vm/client/BlockStoreMockImpl.java
@@ -4,9 +4,7 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.ethereum.vm;
-
-import org.ethereum.vm.client.BlockStore;
+package org.ethereum.vm.client;
 
 public class BlockStoreMockImpl implements BlockStore {
